A bit of pdf formatting + refactoring
